Split Test_nPAClientLib in two files.
